backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Vector Database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.2
torch>=2.1.0
numpy>=1.24.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
